renamed appflow namespace to package
