--- packages/ionic/src/commands/appflow/index.ts
@@ -1,24 +1,0 @@
-import chalk from 'chalk';
-
-import { CommandMap, Namespace } from '../../lib/namespace';
-
-export class AppflowNamespace extends Namespace {
-  async getMetadata() {
-    return {
-      name: 'appflow',
-      summary: 'Appflow functionality',
-      description: `
-Inteface to execute commands on Ionic Appflow.
-
-Appflow documentation:
-- Overview: ${chalk.bold('https://ionicframework.com/docs/appflow/')}
-      `,
-    };
-  }
-
-  async getCommands(): Promise<CommandMap> {
-    return new CommandMap([
-      ['package', async () => { const { PackageCommand } = await import('./package'); return new PackageCommand(this); }],
-    ]);
-  }
-}
